+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Moves the edit signal button in the popup menu button in the app bar.
 - Initializes the calendar with the current day around the middle of the screen.
 - Makes the signal names more compact in the signals calendar screen.
+- Redirects from the signals calendar screen to the signal points screen when clicking on a signal name.
 
 ## 0.16.0