+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.17.0
 
 - Moves the edit signal button in the popup menu button in the app bar.
 - Initializes the calendar with the current day around the middle of the screen.